move some apiHooks into categories file
